+++ src/main/resources/META-INF/services/org.jetbrains.dokka.plugability.DokkaPlugin
@@ -1,1 +1,1 @@
-template.MyAwesomeDokkaPlugin
+template.AnkDokkaPlugin
